fix: Handle error from axios interceptor

+++ src/services/axios.jsx
@@ -27,14 +27,14 @@ instance
   }, function (error) {
     const originalRequest = error.config;
 
-    if (error.response.status === 401 && originalRequest.url === `${backendUrl}/api/token/refresh`) {
+    if (error.response && error.response.status === 401 && originalRequest.url === `${backendUrl}/api/token/refresh`) {
       AuthService.removeToken();
       AuthService.removeRefreshToken();
       window.location.reload();
       return Promise.reject(error);
     }
 
-  if (error.response.status === 401 && !originalRequest._retry) {
+  if (error.response && error.response.status === 401 && !originalRequest._retry) {
     originalRequest._retry = true;
     const refreshToken = CookiesProvider.get("refreshToken");
     return axios.post(`${backendUrl}/api/token/refresh`,
